+++ _test_common/lib/descriptors.dart
@@ -19,14 +19,19 @@ import 'package:test_descriptor/test_descriptor.dart' as d;
 ///
 /// If [versionDependencies] is provided then the keys are the package names
 /// and the values are the exact versions which will be added as a dependency.
-Future<d.FileDescriptor> pubspec(String name,
-    {Iterable<String> currentIsolateDependencies = const [],
-    Map<String, String> pathDependencies = const {},
-    Map<String, String> versionDependencies = const {}}) async {
+///
+/// The [sdkEnvironment] describes the `environment.sdk` field in the pubspec.
+Future<d.FileDescriptor> pubspec(
+  String name, {
+  Iterable<String> currentIsolateDependencies = const [],
+  Map<String, String> pathDependencies = const {},
+  Map<String, String> versionDependencies = const {},
+  String sdkEnvironment = '>=2.12.0 <3.0.0',
+}) async {
   var buffer = StringBuffer()
     ..writeln('name: $name')
     ..writeln('environment:')
-    ..writeln('  sdk: ">=2.9.0 <3.0.0"')
+    ..writeln('  sdk: "$sdkEnvironment"')
     ..writeln('dependencies:');
 
   // Add all deps as `any` deps, real versions are set in dependency_overrides below.