+++ build_vm_compilers/CHANGELOG.md
@@ -1,6 +1,8 @@
 ## 1.0.9-dev
 
 - Update to build version `2.0.0`.
+- Migrate builders to null safety. Note that compiling with sound null safety
+  is not yet supported.
 
 ## 1.0.8